document-manager/microservices/ocr-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pytesseract==0.3.10
numpy==1.24.3
opencv-python==4.8.1.78
python-multipart==0.0.9
Pillow==9.5.0
setuptools>=69.0.0
wheel>=0.42.0 